fix(core): error mutation coming from integration
