+++ src/utils/utils.js
@@ -2,12 +2,12 @@
 import { parse } from 'component-url';
 import get from 'get-value';
 import { v4 as uuid } from '@lukeed/uuid';
-import { LOAD_ORIGIN } from './ScriptLoader';
 import logger from './logUtil';
 import { commonNames } from './integration_cname';
 import { clientToServerNames } from './client_server_name';
 import { CONFIG_URL, RESERVED_KEYS } from './constants';
 import Storage from './storage';
+import { handleError } from './errorHandler';
 
 /**
  *
@@ -115,81 +115,6 @@ function getJSONTrimmed(context, url, writeKey, callback) {
     }
   };
   xhr.send();
-}
-
-/**
- * This function is to add breadcrumbs
- * @param {string} breadcrumb Message to add insight of an user's journey before the error occurred
- */
-function leaveBreadcrumb(breadcrumb) {
-  if (window.rsBugsnagClient) {
-    window.rsBugsnagClient.leaveBreadcrumb(breadcrumb);
-  }
-}
-
-/**
- * This function is to send handled errors to Bugsnag if Bugsnag client is available
- * @param {Error} error Error instance from handled error
- */
-function notifyError(error) {
-  if (window.rsBugsnagClient) {
-    window.rsBugsnagClient.notify(error);
-  }
-}
-
-function handleError(error, customMessage, analyticsInstance) {
-  let errorMessage;
-  try {
-    if (typeof error === 'string') {
-      errorMessage = error;
-    } else if (error instanceof Error) {
-      errorMessage = error.message;
-    } else {
-      errorMessage = error.message ? error.message : JSON.stringify(error);
-    }
-  } catch (e) {
-    errorMessage = '';
-  }
-
-  try {
-    if (error instanceof Event) {
-      // Discard all the non-script loading errors
-      if (error.target && error.target.localName !== 'script') return;
-
-      // Discard script errors that are not originated at SDK or from native SDKs
-      if (
-        error.target.dataset &&
-        (error.target.dataset.loader !== LOAD_ORIGIN ||
-          error.target.dataset.isNonNativeSDK !== 'true')
-      )
-        return;
-
-      errorMessage = `error in script loading:: src::  ${error.target.src} id:: ${error.target.id}`;
-
-      // SDK triggered ad-blocker script
-      if (error.target.id === 'ad-block') {
-        analyticsInstance.page(
-          'RudderJS-Initiated',
-          'ad-block page request',
-          { path: '/ad-blocked', title: errorMessage },
-          analyticsInstance.sendAdblockPageOptions,
-        );
-        // No need to proceed further for Ad-block errors
-        return;
-      }
-    }
-
-    const customeErrMessagePrefix = customMessage || '';
-    errorMessage = `[handleError]::${customeErrMessagePrefix} "${errorMessage}"`;
-    logger.error(errorMessage);
-    let errorObj = error;
-    if (!(error instanceof Error)) errorObj = new Error(errorMessage);
-    notifyError(errorObj);
-  } catch (err) {
-    logger.error('[handleError] Exception:: ', err);
-    logger.error('[handleError] Original error:: ', JSON.stringify(error));
-    notifyError(err);
-  }
 }
 
 function getDefaultPageProperties() {
@@ -756,7 +681,6 @@ export {
   findAllEnabledDestinations,
   transformToRudderNames,
   transformToServerNames,
-  handleError,
   rejectArr,
   type,
   flattenJsonPayload,
@@ -774,8 +698,6 @@ export {
   constructPayload,
   getConfigUrl,
   getSDKUrlInfo,
-  notifyError,
-  leaveBreadcrumb,
   get,
   countDigits,
   getStringId,
